requirements.txt
websockets==12.0
pandas==2.1.3
pyarrow==14.0.1
google-cloud-storage==2.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
